mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "mcp[cli,server]>=1.21.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
